Relevanssi compatibility overhaul
Changed orderby syntax to be compatible with Relevanssi, and added process_orderby() method to convert these values back into a version compatible with regular WP_Query.

Added support for relevanssi_do_query inside of the query() method, so developers won't need to call it separately.

the_form() - When Relevanssi is active, if no orderby field exists, a hidden form field will be generated to specify the "orderby" value.  This is to help ensure that 'orderby' & order' arguments will continue to function when Relevanssi is in use.  Relevanssi does accept "orderby" arguments, although it reads them through the URL rather than the WP_Query object, which is why this field is necessary.

Added support for custom IDs on HTML fields.